requirements.txt
RPi.GPIO
adafruit-circuitpython-adxl34x
